0.9.0

+++ src/main/java/org/cardanofoundation/job/schedules/NativeScriptInfoSchedule.java
@@ -1,26 +1,16 @@
 package org.cardanofoundation.job.schedules;
 
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Set;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.cardanofoundation.explorer.consumercommon.entity.BaseEntity_;
-import org.cardanofoundation.explorer.consumercommon.entity.Script;
-import org.cardanofoundation.explorer.consumercommon.enumeration.ScriptType;
-import org.cardanofoundation.explorer.consumercommon.explorer.entity.NativeScriptInfo;
-import org.cardanofoundation.job.common.enumeration.RedisKey;
-import org.cardanofoundation.job.projection.ScriptNumberHolderProjection;
-import org.cardanofoundation.job.projection.ScriptNumberTokenProjection;
-import org.cardanofoundation.job.repository.explorer.NativeScriptInfoRepository;
-import org.cardanofoundation.job.repository.ledgersync.AddressTokenBalanceRepository;
-import org.cardanofoundation.job.repository.ledgersync.MultiAssetRepository;
-import org.cardanofoundation.job.repository.ledgersync.ScriptRepository;
-import org.cardanofoundation.job.repository.ledgersync.TxRepository;
-import org.cardanofoundation.ledgersync.common.common.nativescript.NativeScript;
-import org.cardanofoundation.ledgersync.common.common.nativescript.RequireTimeAfter;
-import org.cardanofoundation.ledgersync.common.common.nativescript.RequireTimeBefore;
-import org.cardanofoundation.ledgersync.common.common.nativescript.ScriptAll;
-import org.cardanofoundation.ledgersync.common.common.nativescript.ScriptAny;
-import org.cardanofoundation.ledgersync.common.common.nativescript.ScriptAtLeast;
-import org.cardanofoundation.ledgersync.common.common.nativescript.ScriptPubkey;
+
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -31,13 +21,26 @@ import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Set;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.Function;
-import java.util.stream.Collectors;
+import com.bloxbean.cardano.client.transaction.spec.script.NativeScript;
+import com.bloxbean.cardano.client.transaction.spec.script.RequireTimeAfter;
+import com.bloxbean.cardano.client.transaction.spec.script.RequireTimeBefore;
+import com.bloxbean.cardano.client.transaction.spec.script.ScriptAll;
+import com.bloxbean.cardano.client.transaction.spec.script.ScriptAny;
+import com.bloxbean.cardano.client.transaction.spec.script.ScriptAtLeast;
+import com.bloxbean.cardano.client.transaction.spec.script.ScriptPubkey;
+
+import org.cardanofoundation.explorer.common.entity.enumeration.ScriptType;
+import org.cardanofoundation.explorer.common.entity.explorer.NativeScriptInfo;
+import org.cardanofoundation.explorer.common.entity.ledgersync.BaseEntity_;
+import org.cardanofoundation.explorer.common.entity.ledgersync.Script;
+import org.cardanofoundation.job.common.enumeration.RedisKey;
+import org.cardanofoundation.job.projection.ScriptNumberHolderProjection;
+import org.cardanofoundation.job.projection.ScriptNumberTokenProjection;
+import org.cardanofoundation.job.repository.explorer.NativeScriptInfoRepository;
+import org.cardanofoundation.job.repository.ledgersync.AddressTokenBalanceRepository;
+import org.cardanofoundation.job.repository.ledgersync.MultiAssetRepository;
+import org.cardanofoundation.job.repository.ledgersync.ScriptRepository;
+import org.cardanofoundation.job.repository.ledgersync.TxRepository;
 
 @Component
 @RequiredArgsConstructor
@@ -55,7 +58,8 @@ public class NativeScriptInfoSchedule {
   private String network;
 
   private static final int DEFAULT_PAGE_SIZE = 500;
-  private static final List<ScriptType> NATIVE_SCRIPT_TYPES = List.of(ScriptType.TIMELOCK, ScriptType.MULTISIG);
+  private static final List<ScriptType> NATIVE_SCRIPT_TYPES =
+      List.of(ScriptType.TIMELOCK, ScriptType.MULTISIG);
 
   @Scheduled(fixedRateString = "${jobs.native-script-info.fixed-delay}")
   @Transactional(value = "explorerTransactionManager")
@@ -77,7 +81,8 @@ public class NativeScriptInfoSchedule {
     Long txCheckpoint = checkpoint.longValue();
     Set<String> scriptHashList =
         scriptRepository.findHashByTypeAndTxIn(txCheckpoint, currentTxId, NATIVE_SCRIPT_TYPES);
-    scriptHashList.addAll(multiAssetRepository.findPolicyByTxIn(txCheckpoint, currentTxId, NATIVE_SCRIPT_TYPES));
+    scriptHashList.addAll(
+        multiAssetRepository.findPolicyByTxIn(txCheckpoint, currentTxId, NATIVE_SCRIPT_TYPES));
 
     final AtomicInteger counter = new AtomicInteger();
     scriptHashList.stream()
@@ -86,18 +91,19 @@ public class NativeScriptInfoSchedule {
             (page, scriptHashes) -> {
               List<Script> scripts = scriptRepository.findAllByHashIn(scriptHashes);
               saveNativeScriptInfo(scripts);
-            }
-        );
-    log.info("End update NativeScriptInfo with size = {} in {} ms", scriptHashList.size(), System.currentTimeMillis() - startTime);
+            });
+    log.info(
+        "End update NativeScriptInfo with size = {} in {} ms",
+        scriptHashList.size(),
+        System.currentTimeMillis() - startTime);
   }
 
   private void init() {
     log.info("Start init NativeScriptInfo");
     long startTime = System.currentTimeMillis();
-    Pageable pageable = PageRequest.of(0, DEFAULT_PAGE_SIZE, Sort.by(Sort.Direction.ASC, BaseEntity_.ID));
-    Slice<Script> scripts =
-        scriptRepository
-            .findAllByTypeIn(NATIVE_SCRIPT_TYPES, pageable);
+    Pageable pageable =
+        PageRequest.of(0, DEFAULT_PAGE_SIZE, Sort.by(Sort.Direction.ASC, BaseEntity_.ID));
+    Slice<Script> scripts = scriptRepository.findAllByTypeIn(NATIVE_SCRIPT_TYPES, pageable);
     saveNativeScriptInfo(scripts.getContent());
     while (scripts.hasNext()) {
       scripts = scriptRepository.findAllByTypeIn(NATIVE_SCRIPT_TYPES, scripts.nextPageable());
@@ -109,37 +115,43 @@ public class NativeScriptInfoSchedule {
   private void saveNativeScriptInfo(List<Script> scripts) {
     Set<String> scriptHashes = scripts.stream().map(Script::getHash).collect(Collectors.toSet());
     Map<String, NativeScriptInfo> nativeScriptInfoMap =
-        nativeScriptInfoRepository
-            .findAllByScriptHashIn(scriptHashes)
-            .stream()
+        nativeScriptInfoRepository.findAllByScriptHashIn(scriptHashes).stream()
             .collect(Collectors.toMap(NativeScriptInfo::getScriptHash, Function.identity()));
     Map<String, Long> scriptNumberTokenMap =
-        multiAssetRepository
-            .countByPolicyIn(scriptHashes)
-            .stream()
-            .collect(Collectors
-                .toMap(ScriptNumberTokenProjection::getScriptHash, ScriptNumberTokenProjection::getNumberOfTokens));
+        multiAssetRepository.countByPolicyIn(scriptHashes).stream()
+            .collect(
+                Collectors.toMap(
+                    ScriptNumberTokenProjection::getScriptHash,
+                    ScriptNumberTokenProjection::getNumberOfTokens));
     Map<String, Long> scriptNumberHolderMap =
-        addressTokenBalanceRepository.countHolderByPolicyIn(scriptHashes)
-            .stream()
-            .collect(Collectors
-                .toMap(ScriptNumberHolderProjection::getScriptHash, ScriptNumberHolderProjection::getNumberOfHolders));
-    List<NativeScriptInfo> nativeScriptInfoList = scripts.stream().map(item -> {
-      NativeScriptInfo nativeScriptInfo;
-      if (Objects.nonNull(nativeScriptInfoMap.get(item.getHash()))) {
-        nativeScriptInfo = nativeScriptInfoMap.get(item.getHash());
-      } else {
-        nativeScriptInfo = createNativeScriptInfo(item);
-      }
-      nativeScriptInfo.setNumberOfTokens(scriptNumberTokenMap.getOrDefault(item.getHash(), 0L));
-      nativeScriptInfo.setNumberOfAssetHolders(scriptNumberHolderMap.getOrDefault(item.getHash(), 0L));
-      return nativeScriptInfo;
-    }).toList();
+        addressTokenBalanceRepository.countHolderByPolicyIn(scriptHashes).stream()
+            .collect(
+                Collectors.toMap(
+                    ScriptNumberHolderProjection::getScriptHash,
+                    ScriptNumberHolderProjection::getNumberOfHolders));
+    List<NativeScriptInfo> nativeScriptInfoList =
+        scripts.stream()
+            .map(
+                item -> {
+                  NativeScriptInfo nativeScriptInfo;
+                  if (Objects.nonNull(nativeScriptInfoMap.get(item.getHash()))) {
+                    nativeScriptInfo = nativeScriptInfoMap.get(item.getHash());
+                  } else {
+                    nativeScriptInfo = createNativeScriptInfo(item);
+                  }
+                  nativeScriptInfo.setNumberOfTokens(
+                      scriptNumberTokenMap.getOrDefault(item.getHash(), 0L));
+                  nativeScriptInfo.setNumberOfAssetHolders(
+                      scriptNumberHolderMap.getOrDefault(item.getHash(), 0L));
+                  return nativeScriptInfo;
+                })
+            .toList();
     nativeScriptInfoRepository.saveAll(nativeScriptInfoList);
   }
 
   /**
    * create native script info from script
+   *
    * @param script script entity
    * @return native script info
    */
@@ -161,12 +173,12 @@ public class NativeScriptInfoSchedule {
    *
    * @param nativeScript native script
    */
-  public void explainNativeScript(NativeScript nativeScript,
-                                  NativeScriptInfo nativeScriptInfo) {
+  public void explainNativeScript(NativeScript nativeScript, NativeScriptInfo nativeScriptInfo) {
     if (nativeScript.getClass().equals(ScriptPubkey.class)) {
       ScriptPubkey scriptPubkey = (ScriptPubkey) nativeScript;
       if (Objects.nonNull(scriptPubkey.getKeyHash())) {
-        long numberSig = nativeScriptInfo.getNumberSig() == null ? 0L : nativeScriptInfo.getNumberSig();
+        long numberSig =
+            nativeScriptInfo.getNumberSig() == null ? 0L : nativeScriptInfo.getNumberSig();
         nativeScriptInfo.setNumberSig(numberSig + 1);
       }
     } else if (nativeScript.getClass().equals(ScriptAll.class)) {
@@ -186,10 +198,10 @@ public class NativeScriptInfoSchedule {
       }
     } else if (nativeScript.getClass().equals(RequireTimeAfter.class)) {
       RequireTimeAfter requireTimeAfter = (RequireTimeAfter) nativeScript;
-      nativeScriptInfo.setAfterSlot(requireTimeAfter.getSlot().longValue());
+      nativeScriptInfo.setAfterSlot(requireTimeAfter.getSlot());
     } else if (nativeScript.getClass().equals(RequireTimeBefore.class)) {
       RequireTimeBefore requireTimeBefore = (RequireTimeBefore) nativeScript;
-      nativeScriptInfo.setBeforeSlot(requireTimeBefore.getSlot().longValue());
+      nativeScriptInfo.setBeforeSlot(requireTimeBefore.getSlot());
     }
   }
